django_api/pyproject.toml
# ============================
# Projeto
# Referência: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
#
# Instale o pacote: python-dotenv
# Mostrei como no README.md
#
# Os pontos de atenção estão comentados próximo das chaves, aqui só coloquei um
# lembrete.
#
# Atenção: As versões do Python devem ser consistentes entre:
#   - project.requires-python
#   - tool.ruff.target-version
#   - tool.pyright.pythonVersion
#
# Atenção: O nome do pacote principal deve ser consistente entre:
#   - project.scripts
#   - tool.ruff.lint.isort.known-first-party
#   - A pasta dentro de 'src/'
# ============================
[project]
name = "django_api"
version = "0.0.1"
description = "django_api"
readme = "README.md"
license = "MIT"
authors = [{ name = "Victor", email = "victor.ro.dev10@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.1.0",
]

[project.urls]
Homepage = "https://example.org"
# Documentation = "https://example.org"
# Repository = "https://github.com/me/spam.git"
# Issues = "https://github.com/me/spam/issues"
# Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[project.optional-dependencies]
dev = ["ruff", "pyright", "pytest", "pytest-xdist"]

[project.scripts]
# Define os comandos de console. A chave ('django_api') é o comando.
# O valor aponta para 'nome_do_pacote.nome_do_modulo:nome_da_funcao'.
# Ex: 'my_package.main:run' -> src/my_package/main.py (e a função run lá dentro).
# Lembre-se de sincronizar o nome do pacote com 'known-first-party' do Ruff.
django_api = "my_package.my_module:run_from_script"

# ============================
# Lint e formatação (Ruff)
# ============================
[tool.ruff]
line-length = 88
target-version = "py314"
fix = true
show-fixes = true
indent-width = 4
exclude = ["venv", ".venv", "env", ".env", "node_modules", "__pycache__", "**/migrations/**"]

[tool.ruff.lint]
select = [
    "ASYNC",
    "A",
    "ANN",
    "B",
    "BLE",
    "C4",
    "C90",
    "COM",
    "E",
    "EM",
    "ERA",
    "EXE",
    "F",
    "FBT",
    "FIX",
    "I",
    "ICN",
    "ISC",
    "Q",
    "RET",
    "RSE",
    "S",
    "SIM",
    "SLF",
    "T10",
    "T20",
    "TC",
    "TD",
    "TRY",
    "UP",
    "W",
    "YTT",
    "RUF",
    "N",
]
ignore = ["T201", "COM812", "Q000", "E501", "F401"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN201", "S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint.isort]
# Ensina ao Ruff qual é o pacote principal do seu projeto.
# Deve ser o mesmo nome da pasta dentro de 'src/'.
known-first-party = ["my_package"]


# ============================
# Tipagem (Pyright)
# ============================
[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.14"
include = ["src", 'tests']
exclude = [
    "**/venv",
    "**/.venv",
    "**/env",
    "**/.env",
    "**/node_modules",
    "**/__pycache__",
]

reportIncompatibleMethodOverride = false
reportUnreachable = false
venv = ".venv"
venvPath = "."
executionEnvironments = [{ root = "src" }]

# ============================
# Testes (Pytest)
# ============================
[tool.pytest.ini_options]
addopts = "-s --color=yes --tb=short"
# Adiciona 'src' ao path do Python durante os testes. Isso permite que
# os testes em `tests/` importem os módulos de `src/` sem erros.
pythonpath = ["src"]
testpaths = ["tests"]


# ============================
# Build e Setuptools
# ============================
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
# Informa ao setuptools para encontrar os pacotes dentro da pasta 'src'.
# Essencial para a estrutura de layout com 'src/'.
where = ["src"]

[tool.setuptools]
# Mapeia o diretório raiz de pacotes para 'src'.
package-dir = { "" = "src" }
